+++ package.json
@@ -26,7 +26,7 @@
     "gzip-js": "^0.3.2",
     "js-sha256": "^0.9.0",
     "showdown": "^1.9.1",
-    "vue": "^3.2.4",
+    "vue": "^3.2.25",
     "vue-router": "^4.0.11",
     "vuex": "^4.0.2"
   },
