+++ src/components/DnaVersionCard/index.html
@@ -1,0 +1,84 @@
+<div v-if="version" class="card entity-card dna-card">
+    <template v-if="version">
+	<div class="entity-card-header">Version: <strong>{{ name }}</strong></div>
+	<div class="card-body">
+	    <loading :when="loading">
+		<div class="row mb-2">
+		    <div class="col-auto">
+			<p class="card-text text-truncate mb-1">HDK v{{ version.hdk_version }}</p>
+		    </div>
+		    <div class="col text-end">
+			<router-link v-if="link" class="text-decoration-none"
+				     :to="'/dnas/' + parent_id() + '/versions/' + version.$id">
+			    <holo-hash :hash="version.$id" class="float-end ms-3"></holo-hash>
+			</router-link>
+			<holo-hash v-else :hash="version.$id" class="float-end ms-3"></holo-hash>
+		    </div>
+		</div>
+
+		<template v-if="extras">
+		    <h4 class="fs-6 mt-2">Changelog</h4>
+		    <div class="bg-light rounded-3 p-4 pb-3 mb-3">
+			{{ version.changelog || "No changelog" }}
+		    </div>
+		</template>
+
+		<dl v-if="show_parent_ref" class="row mb-0">
+		    <dt class="col-3">DNA</dt>
+		    <dd class="col-9 mb-1" style="margin-left: -0.5em;">
+			<holo-hash :hash="parent_id()"></holo-hash>
+		    </dd>
+		</dl>
+		<dl v-if="extras" class="row mb-0">
+		    <dt class="col-3">Resource Hash</dt>
+		    <dd class="col-9 mb-1">
+			<code>{{ version.wasm_hash }}</code>
+		    </dd>
+		</dl>
+		<dl class="row mb-0">
+		    <dt class="col-3">Created</dt>
+		    <dd class="col-9 mb-1">
+			{{ $filters.time( version.published_at ) }}
+			&mdash; <span class="fw-2">{{ $filters.time( version.published_at, "weekday+date+time" ) }}</span>
+		    </dd>
+		</dl>
+		<dl v-if="extras" class="row mb-0">
+		    <dt class="col-3">Last Updated</dt>
+		    <dd class="col-9 mb-1">
+			{{ $filters.time( version.last_updated ) }}
+			&mdash; <span class="fw-2">{{ $filters.time( version.last_updated, "weekday+date+time" ) }}</span>
+		    </dd>
+		</dl>
+
+		<div v-if="extras" class="row gy-3" style="margin: 1rem 0 2.5rem 0;">
+		    <div v-if="zome_ids.length"
+			 v-for="(zome_version, name) in zomes"
+			 class="col-12 px-0">
+			<zome-version-card v-if="zome_version instanceof Entity"
+					   :title="name"
+					   :link="link"
+					   :entity="zome_version"></zome-version-card>
+			<zome-version-card v-else
+					   :title="name"
+					   :link="link"
+					   :id="zome_version.version"></zome-version-card>
+		    </div>
+		    <div v-else class="card my-4">
+			<div class="card-body text-center p-4">
+			    No zomes
+			</div>
+		    </div>
+		</div>
+
+		<div class="position-absolute bottom-0 end-0">
+		    <a v-show="!extras" @click="toggle_expansion()">
+			<i class="bi-arrows-angle-expand fs-5 float-end flip-x mx-1"></i>
+		    </a>
+		    <a v-show="extras" @click="toggle_expansion()">
+			<i class="bi-arrows-angle-contract fs-5 float-end flip-x mx-1"></i>
+		    </a>
+		</div>
+	    </loading>
+	</div>
+    </template>
+</div>
