+++ src/components/HoloHash/index.html
@@ -1,0 +1,4 @@
+<span class="badge font-monospace"
+      :class="appearance_cls()"
+      :title="hash_str"
+      @dblclick="toggleFullHash()">{{ full_hash ? hash_str : hash_repr }}</span>
