+++ src/components/HoloHash/index.js
@@ -1,0 +1,63 @@
+const { Logger }			= require('@whi/weblogger');
+const log				= new Logger("comp/holo-hash");
+
+const { EntryHash,
+	HeaderHash,
+	DnaHash,
+	AgentPubKey,
+	...HoloHashTypes }		= require('@whi/holo-hash');
+
+
+module.exports = {
+    "props": {
+	"hash": {
+	    "required": true,
+	    validator (value) {
+		if ( value instanceof HoloHashTypes.HoloHash )
+		    return true;
+
+		try {
+		    new HoloHashTypes.HoloHash(value);
+		    return true;
+		} catch (err) {
+		    return false;
+		}
+	    }
+	},
+	"chars": {
+	    "type": Number,
+	    "default": 5,
+	},
+	"expanded": {
+	    "type": Boolean,
+	    "default": false,
+	},
+    },
+    data () {
+	return {
+	    "holohash": new HoloHashTypes.HoloHash( this.hash ),
+	    "hash_str": String( this.hash ),
+	    "full_hash": this.expanded,
+	};
+    },
+    "computed": {
+	hash_repr () {
+	    return this.snip( this.hash_str, 5 );
+	},
+    },
+    "methods": {
+	appearance_cls () {
+	    return {
+		"bg-primary":	this.holohash instanceof AgentPubKey,
+		"bg-light":	this.holohash instanceof EntryHash,
+		"bg-secondary":	this.holohash instanceof HeaderHash,
+		"bg-danger":	this.holohash instanceof DnaHash,
+		"text-dark":	this.holohash instanceof EntryHash,
+	    };
+	},
+	toggleFullHash () {
+	    this.full_hash	= !this.full_hash;
+	},
+    },
+    "template": __template,
+};
