+++ src/components/ListGroup/index.html
@@ -1,0 +1,13 @@
+<div v-if="loading" class="card">
+    <div class="card-body">
+        <loading :when="true"></loading>
+    </div>
+</div>
+<div v-else-if="list.length" class="list-group list-group-flush">
+    <slot></slot>
+</div>
+<div v-else class="card my-4">
+    <div class="card-body text-center">
+        {{ noResultText }}
+    </div>
+</div>
