+++ src/components/ListGroupItem/index.html
@@ -1,0 +1,3 @@
+<div class="list-group-item px-3" :class="{ 'list-group-item-action': this.action }">
+    <slot></slot>
+</div>
