+++ src/components/Loading/index.html
@@ -1,0 +1,6 @@
+<div v-if="when" class="text-center p-4">
+    <div class="spinner-border mt-1" role="status">
+        <span class="visually-hidden">Loading...</span>
+    </div>
+</div>
+<slot v-else></slot>
