+++ src/components/PageView/index.html
@@ -1,0 +1,3 @@
+<div class="flex-grow-1 pt-3 pb-5">
+    <slot></slot>
+</div>
