+++ src/components/PageView/index.js
@@ -1,0 +1,7 @@
+const { Logger }			= require('@whi/weblogger');
+const log				= new Logger("comp/page-view");
+
+
+module.exports = {
+    "template": __template,
+};
