+++ src/components/Placeholder/index.html
@@ -1,0 +1,2 @@
+<slot v-if="!when"></slot>
+<span v-else class="ph-glow" :class="classes" :style="styles"></span>
