+++ src/templates/dnas/single.html
@@ -76,7 +76,7 @@
 			</dt>
 			<dd class="col-sm-12">
 			    <placeholder :when="!$dna.current">
-				<holohash :hash="dna.developer.pubkey"></holohash>
+				<holo-hash :hash="dna.developer.pubkey"></holo-hash>
 			    </placeholder>
 			</dd>
 			<dt class="col-sm-12">
