+++ src/templates/happs/single.html
@@ -87,7 +87,7 @@
 			</dt>
 			<dd class="col-sm-12">
 			    <placeholder :when="!$happ.current">
-				<holohash :hash="happ.designer"></holohash>
+				<holo-hash :hash="happ.designer"></holo-hash>
 			    </placeholder>
 			</dd>
 			<dt class="col-sm-12">
