+++ src/templates/zomes/single.html
@@ -76,7 +76,7 @@
 			</dt>
 			<dd class="col-sm-12">
 			    <placeholder :when="!$zome.current">
-				<holohash :hash="zome.developer.pubkey"></holohash>
+				<holo-hash :hash="zome.developer.pubkey"></holo-hash>
 			    </placeholder>
 			</dd>
 			<dt class="col-sm-12">
